Override toPath() for Arc. Darwin is needed for sin/cos.
